JBRULES-1598 Efficient WorkingMemory serialization with binary protocol -now uses a case statement for the nodes.
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@20151 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ kogito-drools/drools-core/src/main/java/org/drools/marshalling/PersisterEnums.java
@@ -1,39 +1,39 @@
 package org.drools.marshalling;
 
 public class PersisterEnums {
-    public static final int REPEAT                 = 0;
-    public static final int END                    = 1;
-    public static final int FACT_HANDLE            = 2;
-    public static final int LEFT_TUPLE             = 3;
-    public static final int RIGHT_TUPLE            = 4;
+    public static final short REPEAT                 = 0;
+    public static final short END                    = 1;
+    public static final short FACT_HANDLE            = 2;
+    public static final short LEFT_TUPLE             = 3;
+    public static final short RIGHT_TUPLE            = 4;
 
-    public static final int INITIAL_FACT_NODE      = 5;
+    public static final short INITIAL_FACT_NODE      = 5;
     
-    public static final int LEFT_TUPLE_BLOCKED     = 6;
-    public static final int LEFT_TUPLE_NOT_BLOCKED = 7;
+    public static final short LEFT_TUPLE_BLOCKED     = 6;
+    public static final short LEFT_TUPLE_NOT_BLOCKED = 7;
 
-    public static final int ACTIVATION             = 8;
-    public static final int PROPAGATION_CONTEXT    = 9;
+    public static final short ACTIVATION             = 8;
+    public static final short PROPAGATION_CONTEXT    = 9;
     
-    public static final int WORKING_MEMORY_ACTION  = 10;
+    public static final short WORKING_MEMORY_ACTION  = 10;
     
-    public static final int EQUALITY_KEY           = 11;
-    public static final int LOGICAL_DEPENDENCY     = 12; 
+    public static final short EQUALITY_KEY           = 11;
+    public static final short LOGICAL_DEPENDENCY     = 12; 
     
-    public static final int AGENDA_GROUP           = 13;
-    public static final int ACTIVATION_GROUP       = 14;
+    public static final short AGENDA_GROUP           = 13;
+    public static final short ACTIVATION_GROUP       = 14;
     
-    public static final int RULE_FLOW_GROUP        = 15;
-    public static final int RULE_FLOW_NODE         = 16;
+    public static final short RULE_FLOW_GROUP        = 15;
+    public static final short RULE_FLOW_NODE         = 16;
     
-    public static final int PROCESS_INSTANCE          = 17;
-    public static final int NODE_INSTANCE             = 18;
-    public static final int WORK_ITEM                 = 19;
-    public static final int RULE_SET_NODE_INSTANCE    = 20;
-    public static final int WORK_ITEM_NODE_INSTANCE   = 21;
-    public static final int SUB_PROCESS_NODE_INSTANCE = 22;
-    public static final int MILESTONE_NODE_INSTANCE   = 23;
-    public static final int TIMER_NODE_INSTANCE       = 24;
-    public static final int JOIN_NODE_INSTANCE        = 25;
+    public static final short PROCESS_INSTANCE          = 17;
+    public static final short NODE_INSTANCE             = 18;
+    public static final short WORK_ITEM                 = 19;
+    public static final short RULE_SET_NODE_INSTANCE    = 20;
+    public static final short WORK_ITEM_NODE_INSTANCE   = 21;
+    public static final short SUB_PROCESS_NODE_INSTANCE = 22;
+    public static final short MILESTONE_NODE_INSTANCE   = 23;
+    public static final short TIMER_NODE_INSTANCE       = 24;
+    public static final short JOIN_NODE_INSTANCE        = 25;
 
 }
